updates after code review.
